+++ backend/app/api/routes/login.py
@@ -10,7 +10,8 @@ from app.api.deps import CurrentUser, SessionDep, get_current_active_superuser
 from app.core import security
 from app.core.config import settings
 from app.core.security import get_password_hash
-from app.models.user import Message, NewPassword, Token, UserPublic
+from app.models.user import NewPassword, Token, UserPublic
+from app.models.message import Message
 from app.utils import (
     generate_password_reset_token,
     generate_reset_password_email,
