+++ backend/app/models/message.py
@@ -1,0 +1,5 @@
+from sqlmodel import SQLModel
+
+# Generic message
+class Message(SQLModel):
+    message: str
