+++ backend/app/models/user.py
@@ -70,8 +70,3 @@ class TokenPayload(SQLModel):
 class NewPassword(SQLModel):
     token: str
     new_password: str = Field(min_length=8, max_length=40)
-
-# Generic message
-class Message(SQLModel):
-    message: str
-
